[Dy2St] Move transformers to a separate dir
