+++ python/paddle/jit/dy2static/transformers/decorator_transformer.py
@@ -18,8 +18,8 @@ import warnings
 
 from paddle.utils import gast
 
-from .base_transformer import BaseTransformer
-from .utils import RE_PYMODULE, RE_PYNAME, ast_to_source_code
+from ..utils import RE_PYMODULE, RE_PYNAME, ast_to_source_code
+from .base import BaseTransformer
 
 __all__ = []
