+++ python/paddle/jit/dy2static/transformers/utils.py
@@ -14,6 +14,7 @@
 
 
 import ast
+import sys
 
 import astor
 
@@ -31,6 +32,9 @@ def ast_to_source_code(ast_node):
         )
     if isinstance(ast_node, gast.AST):
         ast_node = gast.gast_to_ast(ast_node)
+
+    if sys.version_info >= (3, 9):
+        return ast.unparse(ast_node)
 
     # Do not wrap lines even if they are too long
     def pretty_source(source):
